chore: Slicer shared
