+++ src/containers/TasksProgress/TasksProgress.tsx
@@ -19,26 +19,26 @@ import { setFocusedTasks } from '@state/context'
 import { useAppDispatch } from '@state/store'
 import { useUpdateEntitiesMutation } from '@queries/entity/updateEntity'
 import { toast } from 'react-toastify'
-import { Button, Section, ShortcutTag, Spacer, Toolbar } from '@ynput/ayon-react-components'
+import { Button, Filter, Section, ShortcutTag, Spacer, Toolbar } from '@ynput/ayon-react-components'
 import Shortcuts from '@containers/Shortcuts'
 import { openViewer } from '@state/viewer'
 import EmptyPlaceholder from '@shared/EmptyPlaceholder/EmptyPlaceholder'
 import './styles.scss'
 import { AttributeEnumItem } from '@api/rest/attributes'
-import { Filter } from '@components/SearchFilter/types'
 import SearchFilterWrapper from '@components/SearchFilter/SearchFilterWrapper'
 import formatFilterAttributesData from './helpers/formatFilterAttributesData'
 import formatFilterTagsData from './helpers/formatFilterTagsData'
 import { FilterFieldType } from '@hooks/useBuildFilterOptions'
 import formatFilterAssigneesData from './helpers/formatFilterAssigneesData'
 import { selectProgress } from '@state/progress'
-import { SelectionData, SliceType, useSlicerContext } from '@context/slicerContext'
+import { useSlicerContext } from '@context/slicerContext'
 import useFilterBySlice from './hooks/useFilterBySlice'
 import formatSearchQueryFilters from './helpers/formatSearchQueryFilters'
 import { isEmpty } from 'lodash'
 import { RowSelectionState } from '@tanstack/react-table'
 import { useFiltersWithHierarchy } from '@components/SearchFilter/hooks'
 import useUserFilters from '@hooks/useUserFilters'
+import { SelectionData, SliceType } from '@shared/Slicer'
 
 // what to search by
 const searchFilterTypes: FilterFieldType[] = [
