+++ src/shared/src/index.tsx
@@ -3,3 +3,4 @@ export * from './hooks'
 export * from './helpers'
 export * from './EmptyPlaceholder'
 export * from './ContextMenu'
+export * from './Slicer'
